Create home_load.js
Currently is using scryfall's API to get the images and prices of all the cards. This would be best if it was moved to a module script or something similar in JavaScript and the cardtbl is an argument submitted by another script and the load of the page is then completed, rather than having this script or something similar for each deck and/or other page on the site.